Move check hive table from SelectStmt to FromClause and update doc
