+++ docs/zh-CN/sql-reference/sql-statements/Data Definition/CREATE TABLE.md
@@ -127,10 +127,10 @@ under the License.
 
     注意：
         "table" 条目中的 "table_name" 是 mysql 中的真实表名。
-        而 CREATE TABLE 语句中的 table_name 是该 mysql 表在 Palo 中的名字，可以不同。
+        而 CREATE TABLE 语句中的 table_name 是该 mysql 表在 Doris 中的名字，可以不同。
     
-    在 Palo 创建 mysql 表的目的是可以通过 Palo 访问 mysql 数据库。
-        而 Palo 本身并不维护、存储任何 mysql 数据。
+    在 Doris 创建 mysql 表的目的是可以通过 Doris 访问 mysql 数据库。
+        而 Doris 本身并不维护、存储任何 mysql 数据。
     2) 如果是 broker，表示表的访问需要通过指定的broker, 需要在 properties 提供以下信息：
         ```
         PROPERTIES (
@@ -152,7 +152,7 @@ under the License.
         "path" 中如果有多个文件，用逗号[,]分割。如果文件名中包含逗号，那么使用 %2c 来替代。如果文件名中包含 %，使用 %25 代替
         现在文件内容格式支持CSV，支持GZ，BZ2，LZ4，LZO(LZOP) 压缩格式。
 
-    3) 如果是hive，则需要在properties提供以下信息：
+    3) 如果是 hive，则需要在 properties 提供以下信息：
     ```
     PROPERTIES (
         "database" = "hive_db_name",
@@ -161,8 +161,8 @@ under the License.
     )
 
     ```
-    其中database是hive表对应的库名字，table是hive表的名字，hive.metastore.uris是hive metastore服务地址。
-    注意：目前hive外部表仅用于Spark Load使用。
+    其中 database 是 hive 表对应的库名字，table 是 hive 表的名字，hive.metastore.uris 是 hive metastore 服务地址。
+    注意：目前hive外部表仅用于Spark Load使用，不支持查询。
 
 1. key_desc
     语法：
@@ -174,7 +174,7 @@ under the License.
                              适合报表、多维分析等业务场景。
                 UNIQUE KEY:key列相同的记录，value列按导入顺序进行覆盖，
                              适合按key列进行增删改查的点查询业务。
-                DUPLICATE KEY:key列相同的记录，同时存在于Palo中，
+                DUPLICATE KEY:key列相同的记录，同时存在于Doris中，
                              适合存储明细数据或者数据无聚合特性的业务场景。
         默认为DUPLICATE KEY，key列为列定义中前36个字节, 如果前36个字节的列数小于3，将使用前三列。
     注意：
