+++ fe/src/main/java/org/apache/doris/analysis/FromClause.java
@@ -23,9 +23,15 @@ import java.util.Comparator;
 import java.util.Iterator;
 import java.util.List;
 
+import org.apache.doris.catalog.Database;
+import org.apache.doris.catalog.Table;
+import org.apache.doris.cluster.ClusterNamespace;
 import org.apache.doris.common.AnalysisException;
+import org.apache.doris.common.ErrorCode;
+import org.apache.doris.common.ErrorReport;
 import org.apache.doris.common.UserException;
 
+import com.google.common.base.Strings;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Lists;
 
@@ -72,6 +78,39 @@ public class FromClause implements ParseNode, Iterable<TableRef> {
         });
     }
 
+    private void checkFromHiveTable(Analyzer analyzer) throws AnalysisException {
+        for (TableRef tblRef : tableRefs_) {
+            if (!(tblRef instanceof BaseTableRef)) {
+                continue;
+            }
+
+            TableName tableName = tblRef.getName();
+            String dbName = tableName.getDb();
+            if (Strings.isNullOrEmpty(dbName)) {
+                dbName = analyzer.getDefaultDb();
+            } else {
+                dbName = ClusterNamespace.getFullName(analyzer.getClusterName(), tblRef.getName().getDb());
+            }
+            if (Strings.isNullOrEmpty(dbName)) {
+                ErrorReport.reportAnalysisException(ErrorCode.ERR_NO_DB_ERROR);
+            }
+
+            Database db = analyzer.getCatalog().getDb(dbName);
+            if (db == null) {
+                ErrorReport.reportAnalysisException(ErrorCode.ERR_BAD_DB_ERROR, dbName);
+            }
+
+            String tblName = tableName.getTbl();
+            Table table = db.getTable(tblName);
+            if (table == null) {
+                ErrorReport.reportAnalysisException(ErrorCode.ERR_BAD_TABLE_ERROR, tblName);
+            }
+            if (table.getType() == Table.TableType.HIVE) {
+                throw new AnalysisException("Query from hive table is not supported, table: " + tblName);
+            }
+        }
+    }
+
     @Override
     public void analyze(Analyzer analyzer) throws AnalysisException, UserException {
         if (analyzed_) return;
@@ -97,6 +136,9 @@ public class FromClause implements ParseNode, Iterable<TableRef> {
             tblRef.analyze(analyzer);
             leftTblRef = tblRef;
         }
+
+        // TODO: remove when query from hive table is supported
+        checkFromHiveTable(analyzer);
 
         analyzed_ = true;
     }
